refactor: Border enum으로 변경
move: ItemColor, Rounding 패키지 rules로 이동
