+++ compose/src/main/java/com/yourssu/design/system/compose/atom/Divider.kt
@@ -1,7 +1,16 @@
 package com.yourssu.design.system.compose.atom
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.IntrinsicSize
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -16,8 +25,8 @@ enum class Direction {
 }
 
 enum class Thickness(val value: Dp) {
-    Thin(YdsBorder.thin),
-    Thick(YdsBorder.thick)
+    Thin(YdsBorder.Thin.dp),
+    Thick(YdsBorder.Thick.dp)
 }
 
 @Composable
