+++ compose/src/main/java/com/yourssu/design/system/compose/rule/ItemColor.kt
@@ -1,4 +1,4 @@
-package com.yourssu.design.system.compose.foundation
+package com.yourssu.design.system.compose.rule
 
 import androidx.compose.runtime.Composable
 import com.yourssu.design.system.compose.YdsTheme
